editline: fix non ambiguous completion space
Given the following command tree:

	seq(str("interface"), str("add"), any())

And the following interactive editline session:

	> interface add<tab>

Even though, there is a single non-ambiguous completion for the last
"add" argument, no space is appended to validate it and move the cursor
to begin a new argument.

This occurs because no "append" characters are found (indeed, the last
argument is already valid) and we wrongfully assume that means we have
multiple ambiguous completion choices.

Only assume we have ambiguous completion if there is more than one item.

Signed-off-by: Robin Jarry <robin@jarry.cc>