editline: rework ec_editline_get_error_helps()
Currently, the tokenization is done in a quick and dirty way: when a space is
found in the string, it is replaced by a \0 and it retries to parse/complete to
generate the error helps.

Use the ec_strvec_sh_lex_str() function to enhance the tokenization. It makes
the function more complex, but it can be expected to be more reliable.

Signed-off-by: Olivier Matz <zer0@droids-corp.org>
Reviewed-by: Robin Jarry <robin@jarry.cc>