node_str: fix return value when no completion
When there is no completion in node_str, the function returns
EC_PARSE_NOMATCH, but this return code is not expected from a completion
function. This is harmless, because ec_complete_child() will convert any
positive value into 0.

Fix the return value, and use ec_str_startswith() to simplify the code
and remove a XXX comment.

Fixes: 4b0b6d102274 ("list children in a table, not in a list")
Signed-off-by: Olivier Matz <zer0@droids-corp.org>
Reviewed-by: Robin Jarry <robin@jarry.cc>