editline: add api to set a short description
Add a function to set a short description, displayed in contextual
help. Similar to help and callback mechanisms, it uses a node attribute.

The get_node_help() function that checks for this attribute becomes a
bit more complex, especially because the node_desc is either allocated
or const, depending on where it comes from.

Signed-off-by: Olivier Matz <zer0@droids-corp.org>
Reviewed-by: Robin Jarry <robin@jarry.cc>