editline: print error helps on completion failure
Currently, if we ask for contextual helps (with <?>) on a line that does
not match the grammar, it displays nothing.

It makes more sense to display error helps.

Note: we need to fix ec_editline_print_error_helps() so that it works on
a line that does not end with '\n'.

Signed-off-by: Olivier Matz <zer0@droids-corp.org>
Reviewed-by: Robin Jarry <robin@jarry.cc>