tools: add a tool to generate schema api doc
This tool helps to generate up-to-date documentation. It browses all
registered libecoli nodes types, and for each of them it generates a
section in a markdown document that details its configuration API.

The result of this tool is added in this commit in
doc/node-schema.md. It can be updated with "make node-schema", and a
test checks that documentation is always up to date (so the CI will
check it).

Signed-off-by: Olivier Matz <zer0@droids-corp.org>