node_many: don't break the parsing loop on empty match
Currently, when we try to parse a "many" node, the parsing stops if a
child parses an empty strvec successfully. This was done to avoid an
infinite loop: parsing an empty strvec does not consume anything in the
parent strvec, and the parsing does not make any progress.

However, there are use cases where matching an empty strvec is what we
want. If priv->max is set, there is no risk of infinite loop, so only
break if priv->max == 0.

Signed-off-by: Olivier Matz <zer0@droids-corp.org>