node: introduce grammar graph iterator
A grammar tree is actually not really a tree, it is an oriented graph
where loop can exist. For this reason, it is not possible to iterate the
grammar graph without storing somewhere the list of browsed nodes in
order to break loops.

Another property of grammar graphs is that the same node can appear
several times at different places in the graph, having a different
parent. These kind of nodes will be iterated only once.

Introduce an iterator for grammar graphs, which first builds a real
tree. Each node of this tree references a node of the grammar graph.

The next commit will make use of it for ec_node_find*() functions, that
are currently looping forever if the grammar graph contains a loop.

Signed-off-by: Olivier Matz <zer0@droids-corp.org>